Missed saving the file, so same as previous commit

+++ AnimalShop/AnimalShop/Animal.cs
@@ -7,7 +7,7 @@
         // Get is still public since I see no problem with other parts of the program reading the data
         public int Age { get; protected set; }
         public double Weight { get; protected set; }
-        public float Cost { get; protected set; }
+        public decimal Cost { get; protected set; }
 
         // I added multiple constructors here in order to be able to create generic animals with a set range of values.
         // these values are then sent to the other constructor in order to have one place to set the values in case we want
